Implement step-based side effects
