+++ maestro/swift/Sources/Programs/ProgramStep.swift
@@ -1,4 +1,4 @@
 public protocol ProgramStep {
     var name: String { get }
-    func apply(changes: [LightState], context: StateContext, transition: Double) -> [LightState]
+    func apply(changes: [LightState], effects: [SideEffect], context: StateContext, transition: Double) -> ([LightState], [SideEffect])
 }
